// product-microservice/src/protos/product/CreateProductResponse.ts
// Original file: node_modules/grpc-nest-proto/protos/product.proto


export interface CreateProductResponse {
  'statusCode'?: (number);
  'error'?: (string)[];
  'id'?: (number);
}

export interface CreateProductResponse__Output {
  'statusCode'?: (number);
  'error'?: (string)[];
  'id'?: (number);
}